More refactoring for a better class/package structure
 - move lifecycle classes to new lifecycle package
 - move search resolvers to search package
