+++ modules/core/app/lifecycle/ItemLifecycle.scala
@@ -1,20 +1,19 @@
-package config
+package lifecycle
 
 import defines.EventType
 import models.base.Model
 
 import scala.concurrent.{ExecutionContext, Future}
 
-
 trait ItemLifecycle {
   /**
     * A lifecycle hook called prior to saving an item.
     *
-    * @param id the item ID, if available
-    * @param item the existing item, if available
-    * @param data the data that is about to be saved
+    * @param id    the item ID, if available
+    * @param item  the existing item, if available
+    * @param data  the data that is about to be saved
     * @param event the event type
-    * @param ec an implicit execution context
+    * @param ec    an implicit execution context
     * @tparam MT the type of the model
     * @return a future of (possibly modified) model data
     */
@@ -23,10 +22,10 @@ trait ItemLifecycle {
   /**
     * A lifecycle hook called after saving an item.
     *
-    * @param id the item ID
-    * @param item the saved item
+    * @param id    the item ID
+    * @param item  the saved item
     * @param event the event type
-    * @param ec the implicit execution context
+    * @param ec    the implicit execution context
     * @tparam MT the type of the model
     * @return a future of the model
     */
