+++ modules/core/app/services/search/IdSearchResolver.scala
@@ -1,0 +1,15 @@
+package services.search
+
+import services.data
+import services.data.{ApiUser, DataApi}
+
+import javax.inject.Inject
+import scala.concurrent.{ExecutionContext, Future}
+
+/**
+  * Resolve search hits to DB items by the itemId field
+  */
+case class IdSearchResolver @Inject()(dataApi: DataApi)(implicit executionContext: ExecutionContext) extends SearchItemResolver {
+  def resolve[MT: data.Readable](docs: Seq[SearchHit])(implicit apiUser: ApiUser): Future[Seq[Option[MT]]] =
+    dataApi.withContext(apiUser).fetch(ids = docs.map(_.itemId))
+}
